requirements.txt
# Web framework
flask>=3.0,<4.0
flask-cors>=5.0,<6.0
gunicorn>=23.0,<24.0

# Linguistics
lingpy>=2.6,<3.0
lingrex>=1.4,<2.0
panphon>=0.22,<1.0

# Data
pandas>=2.0,<4.0
pyyaml>=6.0

# ML model (DPD neural reconstruction)
torch>=2.0
pytorch-lightning>=2.0,<3.0
einops
tabulate
torchmetrics>=0.11,<1.0
scikit-learn
scipy
regex
transformers
torchshow


# Testing
pytest>=7.0